requirements.txt
azure-cli >= 2.41.0
azure-devops >= 6.0.0b4